+++ CHANGELOG.md
@@ -7,7 +7,7 @@ the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ---
 
-## [1.4.3] - Unreleased
+## [1.4.3] - 2026-08-29
 
 ### Added